+++ py_bingads/services/negative_keywords.py
@@ -167,12 +167,6 @@ class NegativeKeywords(_base.BingAds):
 
         :returns: A list of IDs of the created negative keywords
         """
-        # TODO: Test
-        if negative_keywords and isinstance(negative_keywords[0], tuple):
-            negative_keywords = [_models.NegativeKeyword(
-                text=kw[0], match_type=kw[1]
-            ) for kw in negative_keywords]
-
         negative_keyword_list = _models.NegativeKeywordList(id=list_id)
         array_of_negative_keyword = _models.ArrayOfNegativeKeyword(
             negative_keywords=negative_keywords